[#1745] Introduce service resource module
