+++ iceoryx2/conformance-tests/src/reader.rs
@@ -23,8 +23,8 @@ pub mod reader {
     use iceoryx2::port::reader::*;
     use iceoryx2::prelude::*;
     use iceoryx2::service::Service;
-    use iceoryx2::service::builder::blackboard::KeyMemory;
     use iceoryx2::service::marker::CustomKeyMarker;
+    use iceoryx2::service::resource::blackboard::KeyMemory;
     use iceoryx2::service::static_config::message_type_details::{TypeDetail, TypeVariant};
     use iceoryx2::testing::generate_service_name;
     use iceoryx2_bb_testing::assert_that;
