+++ iceoryx2/conformance-tests/src/writer.rs
@@ -21,8 +21,8 @@ pub mod writer {
     use iceoryx2::port::writer::*;
     use iceoryx2::prelude::*;
     use iceoryx2::service::Service;
-    use iceoryx2::service::builder::blackboard::KeyMemory;
     use iceoryx2::service::marker::CustomKeyMarker;
+    use iceoryx2::service::resource::blackboard::KeyMemory;
     use iceoryx2::service::static_config::message_type_details::{TypeDetail, TypeVariant};
     use iceoryx2_bb_concurrency::atomic::AtomicU64;
     use iceoryx2_bb_concurrency::atomic::Ordering;
