+++ iceoryx2/src/service/builder/request_response.rs
@@ -26,11 +26,12 @@ use crate::service::builder::{
 use crate::service::dynamic_config::MessagingPatternSettings;
 use crate::service::dynamic_config::request_response::DynamicConfigSettings;
 use crate::service::port_factory::request_response;
+use crate::service::resource::NoResource;
 use crate::service::static_config::StaticConfig;
 use crate::service::static_config::message_type_details::TypeDetail;
 use crate::service::static_config::messaging_pattern::MessagingPattern;
-use crate::service::{NoResource, header, static_config};
 use crate::service::{Service, builder, dynamic_config};
+use crate::service::{header, static_config};
 
 use super::message_type_details::{MessageTypeDetails, TypeVariant};
 use crate::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
