+++ iceoryx2/src/service/port_factory/blackboard.rs
@@ -46,8 +46,8 @@ use crate::constants::MAX_BLACKBOARD_KEY_SIZE;
 use crate::identifiers::UniqueServiceId;
 use crate::node::NodeListFailure;
 use crate::service::attribute::AttributeSet;
-use crate::service::builder::blackboard::{BlackboardResources, KeyMemory};
 use crate::service::marker::CustomKeyMarker;
+use crate::service::resource::blackboard::{BlackboardResources, KeyMemory};
 use crate::service::service_hash::ServiceHash;
 use crate::service::service_name::ServiceName;
 use crate::service::{self, ServiceState, SharedServiceState, dynamic_config, static_config};
