+++ iceoryx2/src/service/port_factory/event.rs
@@ -45,8 +45,9 @@ use super::notifier::PortFactoryNotifier;
 use crate::identifiers::UniqueServiceId;
 use crate::node::NodeListFailure;
 use crate::service::attribute::AttributeSet;
+use crate::service::resource::NoResource;
 use crate::service::service_hash::ServiceHash;
-use crate::service::{self, NoResource, ServiceState, SharedServiceState, static_config};
+use crate::service::{self, ServiceState, SharedServiceState, static_config};
 use crate::service::{ServiceName, dynamic_config};
 use iceoryx2_bb_elementary::CallbackProgression;
 use iceoryx2_bb_elementary_traits::non_null::NonNullCompat;
